优化

+++ DRCategories/Classes/Foundation/NSString+DRExtension.h
@@ -38,7 +38,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (NSString *)phoneFormat;
 
 // 数值格式化，带千分位符，值定最多小数位
-- (NSString *)numberFormatWithMaxDecimalCount:(NSInteger)maxDecimalCount;
+- (NSString *)numberFormatWithMaxDecimalCount:(int)maxDecimalCount;
 
 /// 阿里云图片链接限定图片短边的长度，等比缩放
 /// @param width 短边长度
